Исправлены ошибки в функциях, объеденены модули для создания полноразмерных фото с комментариями в один модуль "Crrate fullsize-photos"
